+++ R/bg_pop_pts.R
@@ -28,11 +28,10 @@
 #' @export
 
 bg_pop_pts <- function(pop_raster_list='', input_locs='', input_raster='', bg_pts=1000, crs="+init=epsg:5070") {
-  if(is.character(input_raster) == F) {
-    raster_surface <- input_raster
-  } else {
-    raster_surface <- raster::raster(input_raster)
-  }
+  raster_surface <- popmaps_prepare_raster(input_raster)$raster
+  input_locs <- popmaps_prepare_locations(input_locs)
+  popmaps_check_whole_number(bg_pts, "`bg_pts`")
+
   cell_size <- raster::res(raster_surface)[1]
   nrows <- raster_surface@nrows
   ncols <- raster_surface@ncols