+++ R/popmap_viz.R
@@ -40,11 +40,9 @@ popmap_viz <- function(pop_raster_list='',input_locs='',input_raster='',maptype=
     stop("The 'viridis' package is required for popmap_viz().", call. = FALSE)
   }
 
-  if(is.character(input_raster) == F) {
-    raster_surface <- input_raster
-  } else {
-    raster_surface <- raster::raster(input_raster)
-  }
+  raster_surface <- popmaps_prepare_raster(input_raster)$raster
+  input_locs <- popmaps_prepare_locations(input_locs)
+
 	cell_size <- raster::res(raster_surface)[1]
 	nrows <- raster_surface@nrows
 	ncols <- raster_surface@ncols